% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{basket_db_funmart}
\alias{basket_db_funmart}
\title{Grocery basket database (144,159 line items)}
\format{
A data frame with 144,159 rows and 4 variables:
\describe{
\item{basket_id}{unique basket id}
\item{order_id}{order id foreign key related to \emph{order_db_funmart}}
\item{product}{product name}
\item{price}{product price}
}
}
\usage{
data(basket_db_funmart)
}
\description{
A dataset containing basket id, order id, products purchased in each basket
and price of products. There were 200 products, with different
probabilities for each, to select from in the fake grocery
store, 'funmart'. Over 140,000 products were bought in all baskets
combined. The customer ids were sampled
from the \emph{order_db_funmart} table (not \emph{customer_db_funmart})
to ensure that only customers that actually placed an order appear
in the \emph{basket_db_funmart} table.
}
\details{
The product names and prices were collected
from \url{https://www.ocado.com} and \url{https://www.elgrocer.com}.
}
\keyword{datasets}
