% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eg_category}
\alias{eg_category}
\title{Product categories available in different grocery stores.}
\format{
A data frame with 3,278 rows and 5 variables:
\describe{
\item{store_name}{name of the store}
\item{category}{category of products}
\item{category_link}{link to the category in a store}
\item{category_image_link}{link to category image}
\item{store_link}{link to the store}
}
}
\source{
\url{https://www.elgrocer.com}
}
\usage{
eg_category
}
\description{
A dataset containing 3,278 product categories in different grocery stores.
}
\keyword{datasets}
