% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eg_data}
\alias{eg_data}
\title{Details for over 15,000 grocery products.}
\format{
A data frame with 15,459 rows and 9 variables:
\describe{
\item{store_name}{name of the store}
\item{category}{category of products}
\item{subcategory}{subcategory of products}
\item{item}{product name}
\item{weight}{product weight}
\item{price}{product price, in GBP}
\item{category_image_link}{link to category image}
\item{item_image_link}{link to product image}
\item{store_link}{link to the store}
}
}
\source{
\url{https://www.elgrocer.com}
}
\usage{
eg_data
}
\description{
A dataset containing the names and other attributes of over 15,000 grocery
products. This table was built by joining \emph{eg_product},
\emph{eg_subcategory} and \emph{eg_category}.
}
\keyword{datasets}
