% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eg_location}
\alias{eg_location}
\title{Names of location areas in UAE}
\format{
A data frame with 131 rows and 2 variables:
\describe{
\item{location}{name of location in UAE}
\item{location_link}{link to view stores in location}
}
}
\source{
\url{https://www.elgrocer.com}
}
\usage{
data(eg_location)
}
\description{
A dataset contianing the names and links of 131 locations that have grocery
stores that provide online delivery services. A location may contain
several stores. The stores may overlap among different locations.
}
\keyword{datasets}
