% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eg_product}
\alias{eg_product}
\title{Details for 15,459 grocery products.}
\format{
A data frame with 15,459 rows and 5 variables:
\describe{
\item{subcategory_link}{link to the subcategory in a store}
\item{item}{product name}
\item{weight}{product weight}
\item{price}{product price, in AED}
\item{item_image_link}{link to product image}
}
}
\source{
\url{https://www.elgrocer.com}
}
\usage{
eg_product
}
\description{
A dataset containing the name, weight, price and image link of more
than 15,000 grocery products.
}
\keyword{datasets}
