% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eg_store}
\alias{eg_store}
\title{Grocery stores that provide online delivery services.}
\format{
A data frame with 184 rows and 11 variables:
\describe{
\item{location}{name of location in UAE}
\item{city}{name of city in UAE}
\item{store_name}{name of the store}
\item{min_order_amount}{minimum order amount to qualify for delivery, in AED}
\item{delivery_within}{time interval for delivery after placing order}
\item{delivery_start}{time delivery service begins for the day}
\item{delivery_end}{time delivery service ends for the day}
\item{delivery_timezone}{timezone}
\item{payment_method}{available payment methods}
\item{store_link}{link to the store}
\item{location_link}{link to view stores in location}
}
}
\source{
\url{https://www.elgrocer.com}
}
\usage{
eg_store
}
\description{
A dataset containing details for 184 grocery stores that provide online
delivery services.
}
\keyword{datasets}
