% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eg_subcategory}
\alias{eg_subcategory}
\title{Product subcategories available in different grocery stores.}
\format{
A data frame with 156 rows and 3 variables:
\describe{
\item{subcategory}{subcategory of products}
\item{subcategory_link}{link to the subcategory in a store}
\item{category_link}{link to the category in a store}
}
}
\source{
\url{https://www.elgrocer.com}
}
\usage{
data(eg_subcategory)
}
\description{
A dataset containing 1,156 product subcategories chosen randomly from 300
categories in different grocery stores.
}
\keyword{datasets}
