% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty_to_na.R
\name{empty_to_na}
\alias{empty_to_na}
\title{Convert empty and null strings to NA}
\usage{
empty_to_na(x)
}
\arguments{
\item{x}{A character string}
}
\value{
NA
}
\description{
Convert empty and null strings to NA
}
\examples{
x <- "this is text"
empty_to_na(x)

y <- ""
empty_to_na(y)

z <- NULL
empty_to_na(z)
}
