% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_energy.R
\name{extract_energy}
\alias{extract_energy}
\alias{extract_kcal}
\title{Extract the energy content of a product from the nutrition table}
\usage{
extract_energy(data, item, nutrition)

extract_kcal(data)
}
\arguments{
\item{data}{Output from \emph{extract_energy} function}

\item{item}{Product name column.}

\item{nutrition}{Nutrition table column.}
}
\value{
A tibble with product name and energy content, if available.

A tibble with product name, energy content (if available)
and energy content converted into kcal unit.
}
\description{
Helper function to extra the energy content (e.g., kcal or kJ) from the
nutrition table of a product that was collected from the
\emph{oc_collect_nutrition_table} function.

Helper function to convert the energy content
(e.g., kcal or kJ) extracted from \emph{extract_energy} function
into kcal unit.
}
\seealso{
\code{\link{oc_collect_nutrition_table}}
}
