% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ingredients.R
\name{extract_ingredients}
\alias{extract_ingredients}
\title{Extract ingredients from data collected with \emph{oc_collect_product_extra}}
\usage{
extract_ingredients(column)
}
\arguments{
\item{column}{The column that contains the ingredients, if available.}
}
\value{
A character string.
}
\description{
Extract ingredients from data collected with \emph{oc_collect_product_extra}
}
