% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_html_elements.R
\name{get_html_element}
\alias{get_html_element}
\alias{get_html_elements}
\title{Find html element(s) using a remote client driver}
\usage{
get_html_element(remDr, ..., type = "text", attribute_selector)

get_html_elements(remDr, ..., type = "text", attribute_selector)
}
\arguments{
\item{remDr}{Remote client driver}

\item{...}{Additional arguments passed on to \emph{rvest::html_elements}. Usually, a \emph{css} or \emph{xpath}.}

\item{type}{Choose 'text', 'attribute', 'attributes' or 'table' as they relate to the \emph{rvest:html_*} functions.}

\item{attribute_selector}{Name of attribute to find, using \emph{rvest::html_attr}}
}
\value{
html element(s)
}
\description{
Find html element(s) using a remote client driver
}
\examples{
\dontrun{
# Initiate server
remDr <- RSelenium::rsDriver(port = netstat::free_port(),
browser = "firefox", verbose = FALSE)$client

# Navigate to website & get html element
remDr$navigate("https://duckduckgo.com")
get_html_element(remDr = remDr, css = ".content-info__title", type = "text")

# Close the server
remDr$close()
gc(remDr)
rm(remDr)
}
}
