% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nytnyt.R
\name{nytnyt}
\alias{nytnyt}
\title{Suspend execution of functions for a random time interval}
\usage{
nytnyt(period = c(1, 2), crayon_col = crayon::green, ...)
}
\arguments{
\item{period}{Lower and upper limits of time interval (seconds) to suspend execution for. Uses \emph{stats::runif} to pick time.}

\item{crayon_col}{Uses the \emph{crayon} package to select color of character string}

\item{...}{Additional vectors or character strings passed to the end of the returned string}
}
\value{
A character string
}
\description{
Suspend execution of functions for a random time interval
}
\examples{
nytnyt(period = c(2, 5), crayon_col = crayon::blue, "An example of a new string\n")
}
