% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oc_category}
\alias{oc_category}
\title{Grocery products categories.}
\format{
A data frame with 13 rows and 2 variables:
\describe{
\item{category}{category of products}
\item{category_link}{link to the category}
}
}
\source{
\url{https://www.ocado.com}
}
\usage{
data(oc_category)
}
\description{
A dataset containing 13 category names and links.
}
\keyword{datasets}
