% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oc_data}
\alias{oc_data}
\title{Details for 8,920 grocery products.}
\format{
A data frame with 8,920 rows and 17 variables:
\describe{
\item{category}{category of products}
\item{brand}{product brand}
\item{product}{product name}
\item{price}{product price}
\item{weight}{product weight}
\item{badge}{organic, vegetarian, freezing, microwave status of product}
\item{shelf_life}{product shelf life time}
\item{country}{product country of origin}
\item{rating}{product rating}
\item{num_of_reviews}{number of reviews for product}
\item{recommend}{percent of customers that recommend product}
\item{ingredient}{product ingredients}
\item{reviews}{text reviews for product}
\item{nutrition}{nutrition table for product}
\item{image_link}{link to product image}
\item{product_link}{link to product}
\item{category_link}{link to category}
}
}
\source{
\url{https://www.ocado.com}
}
\usage{
data(oc_data)
}
\description{
A dataset containing the names and other attributes of almost 9,000 grocery
products. This table was built by joining \emph{oc_product_general},
\emph{oc_product_extra}, \emph{oc_category}, and \emph{oc_product_review}
and \emph{oc_nutrition_table}.
}
\keyword{datasets}
