% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oc_nutrition_table}
\alias{oc_nutrition_table}
\title{Nutrition table of 992 grocery products}
\format{
A data frame with 992 rows and 2 variables:
\describe{
\item{product_link}{link to product}
\item{nutrition}{nutrition table for product}
}
}
\source{
\url{https://www.ocado.com}
}
\usage{
data(oc_nutrition_table)
}
\description{
A dataset containing the nutrition tables for almost 1,000 grocery products.
}
\keyword{datasets}
