% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oc_product_general}
\alias{oc_product_general}
\title{General information for 8,920 grocery products.}
\format{
A data frame with 8,920 rows and 7 variables:
\describe{
\item{product}{product name}
\item{weight}{product weight}
\item{price}{product price}
\item{shelf_life}{product shelf life time}
\item{image_link}{link to product image}
\item{product_link}{link to product}
\item{category_link}{link to category}
}
}
\source{
\url{https://www.ocado.com}
}
\usage{
oc_product_general
}
\description{
A dataset containing the general information for almost 9,000 grocery
products.
}
\keyword{datasets}
