% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oc_product_review}
\alias{oc_product_review}
\title{Reviews for 944 grocery products.}
\format{
A data frame with 944 rows and 2 variables:
\describe{
\item{product_link}{link to product}
\item{reviews}{text reviews for product}
}
}
\source{
\url{https://www.ocado.com}
}
\usage{
data(oc_product_review)
}
\description{
A dataset containing the reviews for almost 1,000 grocery products.
}
\keyword{datasets}
