% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{order_db_funmart}
\alias{order_db_funmart}
\title{Order database (12,500 orders).}
\format{
A data frame with 12,500 rows and 6 variables:
\describe{
\item{order_id}{unique order id}
\item{customer_id}{customer id foreign key related to \emph{customer_db_funmart}}
\item{order_date}{order date}
\item{order_time}{order time}
\item{payment_method}{1 of 3 methods used to pay for the order}
\item{store}{grocery store}
}
}
\usage{
data(order_db_funmart)
}
\description{
A dataset containing order id, customer id, order date, payment method
and order time. 12,500 orders were randomly generated with the help of
the packages \emph{fabricatr} and \emph{wakefield}.
}
\keyword{datasets}
