% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scroll_page.R
\name{scroll_down_and_load}
\alias{scroll_down_and_load}
\alias{scroll_to_top}
\alias{scroll_down_page_perc}
\alias{scroll_up_page_perc}
\alias{click_show_more}
\title{Scroll to dynamically load more of the webpage}
\usage{
scroll_down_and_load(remDr = remDr)

scroll_to_top(remDr = remDr)

scroll_down_page_perc(remDr = remDr, perc = seq(0, 1, 0.005))

scroll_up_page_perc(remDr = remDr, perc = seq(0, 1, 0.0025))

click_show_more(remDr = remDr)
}
\arguments{
\item{remDr}{Remote client driver}

\item{perc}{Incrementally scroll up the page}
}
\value{
A character string
}
\description{
Scroll to dynamically load more of the webpage
}
\examples{
\dontrun{
# Initiate server
remDr <- RSelenium::rsDriver(port = netstat::free_port(),
browser = "firefox", verbose = FALSE)$client

# Navigate to webpage & scroll down incrementally
url <- "https://www.ocado.com/browse"
remDr$navigate(url)
scroll_down_page_perc(remDr = remDr, perc = seq(0, 1, .25))

# Scroll to the top of the webpage
scroll_to_top(remDr = remDr)

# Close the server
remDr$close()
gc(remDr)
rm(remDr)
}
}
