% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_products.R
\name{select_products}
\alias{select_products}
\title{Randomly select products}
\usage{
select_products(
  products,
  probs,
  customer_id,
  min_products,
  mean_products,
  sd_products
)
}
\arguments{
\item{products}{Vector of products to select from.}

\item{probs}{Probabilities of selecting products.}

\item{customer_id}{Vector of customer id's.}

\item{min_products}{Minimum number of products per basket/order.}

\item{mean_products}{Average number of products per basket/order.}

\item{sd_products}{Standard deviation for number of products per basket/order.}
}
\value{
A vector of product names.
}
\description{
Randomly select products to include in customers' baskets, possible with
varying probabilities for different products. This function is meant to
randomize/stimulate a 'grocery shopping experience', assuming the number
of products per basket/order is randomly distributed.
}
\examples{
# 7 products to choose from
items <- c("apple", "orange", "tea", "coffee", "ice cream", "cookie", "jam")

# 3 customers
c_id <- 1:3

# All products are equally likely to be chosen
baskets <- select_products(items, rep(1/7, length(items)),
customer_id = c_id, min_products = 1, mean_products = 3, sd_products = 1)

# Table of customers and their baskets
tibble::tibble(customer_id = c_id, order = baskets) \%>\%
tidyr::separate_rows(order, sep = "@")
}
