% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sound_work_complete.R
\name{sound_work_complete}
\alias{sound_work_complete}
\title{Play 'Work Complete' sound (from \emph{beepr} package) when the 2 expressions are equal}
\usage{
sound_work_complete(expr1, expr2)
}
\arguments{
\item{expr1}{First expression}

\item{expr2}{Second expression to compare to the first expression}
}
\value{
Audio sound & a character string
}
\description{
Play 'Work Complete' sound (from \emph{beepr} package) when the 2 expressions are equal
}
\examples{
# No audio
sound_work_complete(1, 4)
# Audio
sound_work_complete(1, 1)

expr <- 1:10
sound_work_complete(1, length(expr))
sound_work_complete(10, length(expr))
}
