% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_nested_cols.R
\name{trim_nested_cols}
\alias{trim_nested_cols}
\title{Removes whitespace from start and end of every column of every table in a nested tibble}
\usage{
trim_nested_cols(nested_table)
}
\arguments{
\item{nested_table}{A nested tibble.}
}
\value{
The nested table passed on to the argument.
}
\description{
Removes whitespace from start and end of every column of every table in a nested tibble
}
