% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_table.R
\name{unnest_table}
\alias{unnest_table}
\title{Unnest a table from a nested tibble}
\usage{
unnest_table(nested_table, table_name)
}
\arguments{
\item{nested_table}{The nested tibble containing the table to unnest.}

\item{table_name}{A character string indicating the name of the table to unnest.}
}
\value{
A tibble
}
\description{
The name of the column where the table name is stored must be called
'origin', while the name of the column where the table data is stored
must be called 'data'.
}
\examples{
# Create 2 tibbles
fruit <- tibble::tibble(title = c("Apple", "Banana"), price = c(1.12, .98))
store <- tibble::tibble(location = c("Milan", "London"), rating = c(4, 5))
table_list <- list(fruit = fruit, store = store)

# Create nested tibble
nst_tbl <- tibble::enframe(table_list, name = "origin", value = "data")

# Unnest the 'store' tibble
unnest_table(nested_table = nst_tbl, table_name = "store")
}
