% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_length.R
\name{verify_grocer_length_match}
\alias{verify_grocer_length_match}
\alias{verify_oc_length_match}
\title{Verify that elgrocer output length from selenium & rvest match - 2 objects}
\usage{
verify_grocer_length_match(sel, rve)

verify_oc_length_match(obj1, obj2, obj3, obj4, obj5)
}
\arguments{
\item{sel}{selenium output length}

\item{rve}{rvest output length}

\item{obj1}{titles output length}

\item{obj2}{weights output length}

\item{obj3}{prices output length}

\item{obj4}{images output length}

\item{obj5}{links output length}
}
\value{
logical

logical
}
\description{
Verify that elgrocer output length from selenium & rvest match - 2 objects
}
