% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{customer_db_funmart}
\alias{customer_db_funmart}
\title{Customer database (4,996 customers).}
\format{
A data frame with 4,996 rows and 6 variables:
\describe{
\item{customer_id}{unique customer id}
\item{customer_name}{customer name generated via \emph{randomNames}}
\item{customer_age}{customer age generated via \emph{wakefield}, 18 to 75}
\item{household_size}{household size generated via \emph{wakefield}, 1 to 7}
\item{long}{longitude of customers' delivery address in the UAE
generated via \emph{charlatan}}
\item{lat}{latitude of customers' delivery address in the UAE
generated via \emph{charlatan}}
}
}
\usage{
data(customer_db_funmart)
}
\description{
A dataset containing customer id, name, age, household size and location.
Almost 5,000 customer entries were randomly generated with the help of
the packages \emph{charlatan}, \emph{fabricatr}, \emph{randomNames}
and \emph{wakefield}.
}
\keyword{datasets}
