% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grocer_funcs.R
\name{eg_collect_location_links}
\alias{eg_collect_location_links}
\alias{eg_collect_stores_details}
\alias{eg_collect_categories}
\alias{eg_collect_subcategories}
\alias{eg_collect_items}
\title{Collect elgrocer data}
\usage{
eg_collect_location_links(remDr = remDr, url = "https://www.elgrocer.com")

eg_collect_stores_details(
  remDr = remDr,
  links_to_use,
  sleep_min = 0,
  sleep_max = 1,
  url = "https://www.elgrocer.com"
)

eg_collect_categories(
  remDr = remDr,
  links_to_use,
  sleep_min = 0,
  sleep_max = 1,
  url = "https://www.elgrocer.com"
)

eg_collect_subcategories(
  remDr = remDr,
  links_to_use,
  sleep_min = 0,
  sleep_max = 1,
  url = "https://www.elgrocer.com"
)

eg_collect_items(remDr, links_to_use, sleep_min = 0, sleep_max = 1)
}
\arguments{
\item{remDr}{Remote client driver}

\item{url}{elgrocer url}

\item{links_to_use}{Subcategory links}

\item{sleep_min}{Minimum time to suspend executing R expressions}

\item{sleep_max}{Maximum time to suspend executing R expressions}
}
\value{
Tibble with the URL for each location

Tibble with store links

Tibble with category links

Tibble with subcategory links

Tibble with product details
}
\description{
Collect elgrocer data
}
\examples{
# Initiate server
remDr <- RSelenium::rsDriver(port = 4446L,
browser = "firefox", verbose = FALSE)$client

# Collect all location links
grocer_location <- eg_collect_location_links(remDr = remDr, url = "https://www.elgrocer.com")

# Collect store details from 5 locations
grocer_store <- eg_collect_stores_details(remDr, grocer_location$location_link[1:5])

# Collect categories from 3 stores
grocer_category <- eg_collect_categories(remDr, grocer_store$store_link[1:3])

# Collect subcategories from 3 categories
random_category_links <- sample(1:length(grocer_category$category_link),
3, replace = FALSE)
grocer_subcategory <- eg_collect_subcategories(remDr = remDr, links_to_use = grocer_category$category_link[random_category_links])

# Collect product data from 4 subcategories
random_subcategory_links <- sample(1:length(grocer_subcategory$subcategory_link),
4, replace = FALSE)
grocer_item <- eg_collect_items(grocer_subcategory$subcategory_link[random_subcategory_links])

# Close the server
remDr$close()
gc(remDr)
rm(remDr)
}
