% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{order_db_funmart}
\alias{order_db_funmart}
\title{Order database (12,000 orders)}
\format{
A data frame with 12,500 rows and 6 variables:
\describe{
\item{order_id}{unique order id}
\item{customer_id}{customer id foreign key related to \emph{customer_db_funmart}}
\item{order_date}{order date}
\item{order_time}{order time}
\item{payment_method}{1 of 3 methods used to pay for the order}
\item{store}{grocery store}
}
}
\usage{
data(order_db_funmart)
}
\description{
A dataset containing order id, customer id, order date, payment method
and order time. 12,000 orders were randomly generated with the help of
the packages \emph{fabricatr} and \emph{wakefield}. Not every customer
from the \emph{customer_db_funmart} table might have placed an order.
}
\details{
For example, some customers might have signed up (i.e., input their name,
email address, age, location, etc) and browsed the app, causing
them to be added to the customers' database, but they never actually
completed checkout, so there is no order with their customer id in
the \emph{order_db_funmart} table.
}
\keyword{datasets}
